% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tracking_video.R
\name{create_tracking_video}
\alias{create_tracking_video}
\title{Create a representative video}
\usage{
create_tracking_video(tracking, in_path, out_path, fps = 5, duration = 60,
  start_frame = 1)
}
\arguments{
\item{tracking}{The output of "track"}

\item{in_path}{The path of the original video file}

\item{out_path}{The path of the output video file}

\item{fps}{The "fps" value used when loading the video for tracking}

\item{duration}{The "duration" value used when loading the video for tracking}

\item{start_frame}{The "start_frame" value used when loading the videov}
}
\description{
This function creates a video with the tracking data created with "track"
overlayed on corresponding frames of the raw video used for tracking
}
