% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_tracking.R
\name{inspect_tracking}
\alias{inspect_tracking}
\title{Plot the tracking result}
\usage{
inspect_tracking(tracking, ...)
}
\arguments{
\item{tracking}{The output of "track"}

\item{...}{Additional parameters passed to "plot"}
}
\description{
This function plots the tracking data created with "track"
}
