% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_video.R
\name{load_video}
\alias{load_video}
\title{Load a video}
\usage{
load_video(im_path, fps = 5, duration = 60, start_frame = 1)
}
\arguments{
\item{im_path}{Path to the input file}

\item{fps}{Desired frames per second}

\item{duration}{Length of the video segment in seconds}

\item{start_frame}{The first frame to import}
}
\value{
An image with the extracted frames along the "z" coordinates
}
\description{
This function loads a video file
}
