% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tracking.R
\name{save_tracking}
\alias{save_tracking}
\title{Save the tracking result}
\usage{
save_tracking(path, tracking)
}
\arguments{
\item{path}{The path of the output csv file}

\item{tracking}{The output of "track"}
}
\description{
This function save the tracking data created with "track" as a csv file
}
