% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_mask.R
\name{select_mask}
\alias{select_mask}
\title{Create a mask}
\usage{
select_mask(video, frame = 1, threshold = 0.3, dilation = 100)
}
\arguments{
\item{video}{An image with the extracted frames along the "z" coordinates}

\item{frame}{The frame number to use as the basis for the mask}

\item{threshold}{The luminance threshold, 0:1, used to create the mask}

\item{dilation}{The thickness of the band of mask edge pixels to crop}
}
\value{
A mask matrix valued 0 inside the mask
}
\description{
Create a mask for the tracking area
}
