% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Track the mouse's position}
\usage{
track(video, mask, smoothing = F)
}
\arguments{
\item{video}{An image with the extracted frames along the "z" coordinates}

\item{mask}{A mask matrix made with create_mask}

\item{smoothing}{if TRUE, smooth the tracking result with a rolling median filter}
}
\value{
A dataframe with the mouse's centroid in x and y, in pixels, at each time and frame
}
\description{
This function tracks the mouse's position at all frames in the video
}
